+++ src/state/state.hpp
@@ -5,10 +5,10 @@
 
 template <int Dim> struct State {
   const Map<Dim> map_;
-  const std::vector<Eigen::Transform<float, Dim, Eigen::Affine>>
+  const std::vector<Eigen::Transform<double, Dim, Eigen::Affine>>
       transformations_;
 
   State(const Map<Dim> &map,
-        const std::vector<Eigen::Transform<float, Dim, Eigen::Affine>>
+        const std::vector<Eigen::Transform<double, Dim, Eigen::Affine>>
             &transformations);
 };
